display sites in a grid and button to navigate back
Reading the site_type jsons from db to get the pins details and saving along with device jsons.

TODO - identify the right folder structure to save the site_types data and make it available in application resource

Added a couple of examples for testing locally

Replaced the grid-view content with sub grid sites data.

Modified the view sites button to function as back button which navigates back to the grid-view.

Since the url is not modified in sites view, reloading the window to display tile view with the previous grid selected on clicking back button.